rename: 유저 정보 관련 타입 위치 수정
