+++ src/layouts/RootLayout.tsx
@@ -1,13 +1,10 @@
 import GNB from '@/components/Home/GNB';
 import { Outlet } from 'react-router-dom';
 import { useGetUserProfile } from '@/apis/mypage/getUserProfile';
-import { hasAuthTokens } from '@/utils/authStorage';
 
 const RootLayout = () => {
   // 토큰이 있을 때만 유저 정보 자동 조회
-  useGetUserProfile({
-    enabled: hasAuthTokens(),
-  });
+  useGetUserProfile();
 
   return (
     <main className="mx-auto min-w-1440 max-w-1920 w-full min-h-screen">
